Update: for SOLID principles
